Rename code_800B5F0 to reg_control
