+++ src/main.c
@@ -1,13 +1,13 @@
 #include "global.h"
 #include "bg_palette_buffer.h"
 #include "cpu.h"
-#include "code_800B5F0.h"
 #include "crt0.h"
 #include "file_system.h"
 #include "flash.h"
 #include "input.h"
 #include "main.h"
 #include "random.h"
+#include "reg_control.h"
 #include "sprite.h"
 #include "text1.h"
 #include "text2.h"
